refactor: standard Go CLI layout with cmd/internal split (CLI-002)
