+++ cli/cmd/dot/main_test.go
@@ -1,0 +1,9 @@
+package main
+
+import "testing"
+
+func TestVersionDefault(t *testing.T) {
+	if version == "" {
+		t.Fatal("version must never be empty; default to \"dev\"")
+	}
+}
